Chore/fix stream data

+++ src/database/migrations/20230302144401-transaction-history.js
@@ -19,10 +19,10 @@ module.exports = {
         onUpdate: 'CASCADE'
       },
       amount: {
-        type: Sequelize.DECIMAL
+        type: Sequelize.BIGINT
       },
       credit: {
-        type: Sequelize.DECIMAL
+        type: Sequelize.BIGINT
       },
       createdAt: {
         type: Sequelize.BIGINT,
